pyproject.toml
[tool.poetry]
name = "blog-flask"
version = "0.1.0"
description = "My flask blog"
authors = ["CannedKilroy <GenericArt@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
flask-wtf = "^1.2.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-login = "^0.6.3"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
